Fix empty get_citedby() result
Referring to this issue: https://github.com/OrganicIrradiation/scholarly/issues/47

In the Publisher init method, I extracted the list of cited publications from the URL and pushed it into id_scholarcitedby.

After that, I also had to edit the get_citedby method class.
I had to transform the list of strings in a format that is acceptable from the requests lib. To do that, I had to import "urllib" library too.