Adding a scheduled build every night
We want to get notified if some change on Google Scholar breaks the code. So we will run the build every night to detect breaking changes. There is the chance that we may get a false alert if the test suite breaks due to some network / captcha / etc. issue, but better having false alerts rather than having a broken package.